streamlit/requirements.txt
streamlit
plotly
altair
psycopg2-binary
sqlalchemy
pandas
numpy
openpyxl
python-docx
python-dotenv
requests
streamlit-option-menu
streamlit-aggrid
streamlit-plotly-events
streamlit-chat